Revisit QR code image generation in controller
Remove redundant request parameter decoding, Spring's `@RequestParam`
does this out of the box. But because the web client's `uri()` function,
when given a string encodes that string by default it appeared as if we
have to decode the request parameter (in reality though it was double
encoded, that's why). Using `URI.create()` now in tests to make the uri
being passed explicit.

Refactored image generation so that it forms a single, reactive
pipeline. This way all of the computation becomes lazy/async. Before
this change the image generation in the controller was potentially
blocking, only after the image was created a Mono was populated and
returned. Blocking code in the controller should be avoided, let the
framework be in control when to get the result of a computation.