NO-TICKET: Configure code coverage reports for test suites
The plugin jacoco-report-aggregation allows aggregation of reports for
the same type of tests. We could declare the integration tests and unit
tests to have the same type - but probably we will learn more from
separate reports.
The exlusion of ApplicationKt has been removed for simplicity of
configuration, and because we should not care too much about the coverage
percentage itself. A high coverage is not a goal, just an indicator.